+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-### Fixed
-`src/LockManager.php` - component init before user authentification
+## 1.0.0-alpha2 - 2017-08-18
 
 ### Added
-`src/rest/Action.php` - checkAccess call
+- `Bootstrap` - bootstrapping component as well as an localization initializator.
+
+### Fixed
+- `src/LockManager.php` - component init before user authentification.
+- `checkAccess` callable in `notamedia\locker\rest\Action`.
+
+## 1.0.0-alpha - 2017-07-28